+++ docs/CHANGELOG.md
@@ -10,6 +10,8 @@
 * Added `LoadedExecutable.Execute.OnDeviceNum` to allow  specification of the device by device number in the addressable devices list.
 * Removed the awkward `pjrt.FlatDataToRawWithDimensions` and added the more ergonomic `Client.BufferFromHost.FromFlatDataWithDimensions`.
 * Added `Buffer.ToFlatDataAndDimensions`
+* Store client link with Buffer. Added `Buffer.Client` method.
+* Added `Buffer.Device` and `Client.NumForDevice`.
 
 # v0.1.2 SuppressAbseilLoggingHack